Consolidated common DAO logic in AbstractJdbcDao, and created PhoneDao and implemented most of its methods.

+++ server/src/test/java/io/craigmiller160/orgbuilder/server/data/jdbc/MemberDaoTest.java
@@ -96,7 +96,7 @@ public class MemberDaoTest {
     @Test
     public void testCount() throws Exception{
         testInsert();
-        int count = memberDao.getCount();
+        long count = memberDao.getCount();
         assertEquals("Incorrect count of members in database", 1, count);
     }
